% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn_args_addin.R
\name{get_fn_args}
\alias{get_fn_args}
\title{get_fn_args}
\usage{
get_fn_args(string_select = NULL, valid_only = TRUE,
  return_string = FALSE)
}
\arguments{
\item{string_select}{String to be read and given to the console}

\item{valid_only}{Will only output code which uses = or <-}

\item{return_string}{If True, returns a string of arguments, otherwise outputs directly to console. False by default}
}
\value{
The console will run the provided string split by ; where commas are, or return the string if return_string = T
}
\description{
This function takes a string of arguments as if they were written for a function
it replaces commas with ; where appropriate and executes the code
Designed to easily grab several arguments from a selection.
Note that this works by looking for matching brackets and inverted commas
If for some reason you have hanging commas and speech marks nested within each other this could cause issues!
}
