% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_open_close.R
\name{remove_open_brackets}
\alias{remove_open_brackets}
\title{remove_open_brackets}
\usage{
remove_open_brackets(stringin)
}
\arguments{
\item{stringin}{string to remove brackets from}
}
\description{
Function to remove opening brackets
}
