% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn_args_addin.R
\name{speech_handler}
\alias{speech_handler}
\title{speech_handler}
\usage{
speech_handler(curval, inval, speech)
}
\arguments{
\item{curval}{the current character value}

\item{inval}{current value}

\item{speech}{the type of speech mark being looked for}
}
\value{
inval a numeric vector of length 1
}
\description{
Convenience function to handle speech marks
}
