% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn_args_addin.R
\name{str_iterate}
\alias{str_iterate}
\title{str_iterate}
\usage{
str_iterate(stringin, location = 1, bracket_count = 0, curlbrackets = 0,
  singlespeech = 0, doublespeech = 0, comma_loc = c())
}
\arguments{
\item{stringin}{string to explore. Note that this is the only argument that should normally be provided}

\item{location}{starting location to explore. Should normally be set to 1}

\item{bracket_count}{number of open brackets}

\item{curlbrackets}{number of open curly brackets}

\item{singlespeech}{number of open single speech marks}

\item{doublespeech}{number of open double speech marks}

\item{comma_loc}{vector containing the location of all commas}
}
\value{
comma_loc, a numeric vector with the location of commas in the inputted vector
}
\description{
recursive function for going through a string and finding commas which indicate new arguments
will not locate commas if they are contained by (),{},'' or ""
will ignore escaped characters for this purpose
}
