% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn_args_addin.R
\name{validate_code}
\alias{validate_code}
\title{validate_code}
\usage{
validate_code(stringin)
}
\arguments{
\item{stringin}{A string with R code to be evaulated}
}
\value{
a string with invalid code excluded
}
\description{
Simple function which looks at code segments and sees if theres a definition
That is, currently <-, = or assign.
}
